~ disabled SET global max_allowed_packet for now ~ fixed role value not being properly read when the number exceeds limits of signed long - removed some warnings ~ improved service manager call detection (PyNone are now passed as null to whatever type is specified) + added marketProxy + added calls for the Wallet window (GetJournal, GetEntryTypes, GetSharesByShareholder, CharGetNewTransactions, GetCharOrders) ~ changed market_transactions table

+++ Server/ClusterController/Database/AccountDB.cs
@@ -79,7 +79,7 @@ namespace ClusterControler.Database
 
                     accountid = reader.GetInt64(0);
                     banned = reader.GetBoolean(2);
-                    role = reader.GetInt64(3);
+                    role = (long) reader.GetUInt64(3);
 
                     return true;
                 }
